+++ CHANGES.md
@@ -3,8 +3,8 @@
 ### Added
 
 - **irmin-pack**
-   - Added `integrity-check-index` command in `irmin-fsck`. (#1480, @icristescu,
-     @samoht)
+   - Added `integrity-check-index` command in `irmin-fsck`. (#1480, #1487
+     @icristescu, @samoht)
 
 ### Changed